+++ contracts/zero-ex/test/wrappers.ts
@@ -99,6 +99,7 @@ export * from '../test/generated-wrappers/meta_transactions_feature';
 export * from '../test/generated-wrappers/mixin_aave_v2';
 export * from '../test/generated-wrappers/mixin_balancer';
 export * from '../test/generated-wrappers/mixin_balancer_v2';
+export * from '../test/generated-wrappers/mixin_balancer_v2_batch';
 export * from '../test/generated-wrappers/mixin_bancor';
 export * from '../test/generated-wrappers/mixin_co_fi_x';
 export * from '../test/generated-wrappers/mixin_compound';