+++ packages/asset-swapper/src/utils/market_operation_utils/types.ts
@@ -210,6 +210,23 @@ export interface CurveFillData extends FillData {
     pool: CurveInfo;
 }
 
+export interface BalancerBatchSwapStep {
+    poolId: string;
+    assetInIndex: number;
+    assetOutIndex: number;
+    amount: BigNumber;
+    userData: string;
+}
+
+export interface BalancerSwaps {
+    swapInfoExactIn: BalancerSwapInfo[];
+    swapInfoExactOut: BalancerSwapInfo[];
+}
+export interface BalancerSwapInfo {
+    assets: string[];
+    swapSteps: BalancerBatchSwapStep[];
+}
+
 export interface BalancerFillData extends FillData {
     poolAddress: string;
 }
@@ -217,6 +234,12 @@ export interface BalancerFillData extends FillData {
 export interface BalancerV2FillData extends FillData {
     vault: string;
     poolId: string;
+}
+
+export interface BalancerV2BatchSwapFillData extends FillData {
+    vault: string;
+    swapSteps: BalancerBatchSwapStep[];
+    assets: string[];
 }
 
 export interface UniswapV2FillData extends FillData {